Refact assa effects
