+++ src/UI/Features/Assa/AssaApplyAdvancedEffect/Effects/AdvancedEffectEndCreditsScroll.cs
@@ -1,0 +1,66 @@
+﻿using Nikse.SubtitleEdit.Core.Common;
+using Nikse.SubtitleEdit.Features.Main;
+using Nikse.SubtitleEdit.Logic.Config;
+using Nikse.SubtitleEdit.Logic.Media;
+using System;
+using System.Collections.Generic;
+
+namespace Nikse.SubtitleEdit.Features.Assa.AssaApplyAdvancedEffect.Effects;
+
+public class AdvancedEffectEndCreditsScroll : IAdvancedEffectDisplay
+{
+    public string Name => Se.Language.Assa.AdvancedEffectEndCreditsScroll;
+    public string Description => Se.Language.Assa.AdvancedEffectEndCreditsScrollDescription;
+    public bool UsesAudio => false;
+
+    public override string ToString() => Name;
+
+    public List<SubtitleLineViewModel> ApplyEffect(string header, List<SubtitleLineViewModel> subtitles, int width, int height, WavePeakData2? wavePeaks)
+    {
+        var result = new List<SubtitleLineViewModel>();
+        if (subtitles.Count == 0) return result;
+
+        int screenWidth = width > 0 ? width : 1356;
+        int screenHeight = height > 0 ? height : 678;
+        int centerX = screenWidth / 2;
+
+        double travelDurationMs = 12000;
+        double leadTimeMs = 500;
+        int internalLineSpacing = 60; // Slightly increased for the bigger font
+
+        int yStartBase = screenHeight + 20;
+        int yEndBase = -150;
+
+        foreach (var sub in subtitles)
+        {
+            string processedText = sub.Text.Replace("\\N", "\n").Replace("\\n", "\n");
+            var clean = Utilities.RemoveSsaTags(processedText);
+            if (string.IsNullOrWhiteSpace(clean)) continue;
+
+            var lines = clean.Split(new[] { "\n" }, StringSplitOptions.RemoveEmptyEntries);
+
+            for (int i = 0; i < lines.Length; i++)
+            {
+                var charLine = new SubtitleLineViewModel(sub, generateNewId: true);
+
+                double startTimeMs = sub.StartTime.TotalMilliseconds - leadTimeMs;
+                if (startTimeMs < 0) startTimeMs = 0;
+
+                charLine.StartTime = TimeSpan.FromMilliseconds(startTimeMs);
+                charLine.EndTime = charLine.StartTime.Add(TimeSpan.FromMilliseconds(travelDurationMs));
+
+                int yOffset = i * internalLineSpacing;
+
+                string tags =
+                    $@"\an8\fad(300,300)" +
+                    $@"\move({centerX},{yStartBase + yOffset},{centerX},{yEndBase + yOffset},0,{(int)travelDurationMs})";
+
+                charLine.Text = "{" + tags + "}" + lines[i].Trim();
+                result.Add(charLine);
+            }
+        }
+
+        return result;
+    }
+}
+
